# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DEMO",
        "description": "API para gestionar catgorías, doctores, hospitales y otros recursos.",
        "contact": {
            "email": "esnider@codestation.pe"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://apifap.codestation.pe/api-demo",
            "description": "Servidor prod"
        },
        {
            "url": "http://swagger.local",
            "description": "Servidor local"
        },
        {
            "url": "http://swagger.local",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/hospital": {
            "get": {
                "tags": [
                    "Hospitales"
                ],
                "summary": "Obtiene la lista de hospitales",
                "operationId": "8a8ce0effa77b10be0c75e10f5afde1f",
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID del hospital",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Nombre del hospital",
                                                "type": "string"
                                            },
                                            "address": {
                                                "description": "Dirección del hospital",
                                                "type": "string"
                                            },
                                            "phone_number": {
                                                "description": "Teléfono del hospital",
                                                "type": "string"
                                            },
                                            "email": {
                                                "description": "Correo del hospital",
                                                "type": "string"
                                            },
                                            "deleted": {
                                                "description": "Estado de eliminación lógica",
                                                "type": "boolean"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-02-23T00:09:16.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-02-23T12:33:45.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hospital no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Hospital] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Hospitales"
                ],
                "summary": "Crear un nuevo hospital",
                "description": "Esta operación permite crear un nuevo registro de hospital en la base de datos.",
                "operationId": "d8abc9b3c198047653912054244f7c03",
                "requestBody": {
                    "description": "Datos necesarios para crear un hospital",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "phone_number",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre del hospital",
                                        "type": "string",
                                        "example": "Hospital General"
                                    },
                                    "address": {
                                        "description": "Dirección del hospital",
                                        "type": "string",
                                        "example": "Av. Siempre Viva 123"
                                    },
                                    "phone_number": {
                                        "description": "Teléfono del hospital",
                                        "type": "string",
                                        "example": "555-1234"
                                    },
                                    "email": {
                                        "description": "Correo electrónico del hospital",
                                        "type": "string",
                                        "example": "contacto@hospital.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hospital creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID del hospital",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Nombre del hospital",
                                                    "type": "string"
                                                },
                                                "address": {
                                                    "description": "Dirección del hospital",
                                                    "type": "string"
                                                },
                                                "phone_number": {
                                                    "description": "Teléfono del hospital",
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "description": "Correo electrónico del hospital",
                                                    "type": "string"
                                                },
                                                "deleted": {
                                                    "description": "Estado de eliminación lógica",
                                                    "type": "boolean"
                                                },
                                                "created_at": {
                                                    "description": "Fecha de creación",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "description": "Última fecha de actualización",
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "state": {
                                            "description": "Estado de la operación",
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en el servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "state": {
                                            "description": "Estado de la operación",
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "error_message": {
                                            "description": "Mensaje de error",
                                            "type": "string",
                                            "example": "Mensaje de error detallado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}